+++ app/src/main/java/com/team23/facts23/presentation/views/SnackBarHost.kt
@@ -2,6 +2,7 @@ package com.team23.facts23.presentation.views
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CornerSize
 import androidx.compose.material.*
@@ -12,10 +13,16 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
 import com.team23.achievements.presentation.viewobjects.AchievementPreviewVO
+import com.team23.core.domain.Screen
 
 @Composable
-fun SnackbarHost(achievementPreview: AchievementPreviewVO, hostState: SnackbarHostState) {
+fun SnackbarHost(
+    achievementPreview: AchievementPreviewVO,
+    hostState: SnackbarHostState,
+    navController: NavHostController
+) {
     SnackbarHost(
         hostState = hostState,
         snackbar = {
@@ -23,6 +30,9 @@ fun SnackbarHost(achievementPreview: AchievementPreviewVO, hostState: SnackbarHo
                 elevation = 10.dp, modifier = Modifier
                     .padding(8.dp)
                     .clip(shape = MaterialTheme.shapes.large)
+                    .clickable(onClick = {
+                        navController.navigate(Screen.Achievement().route)
+                    })
             ) {
                 Row(
                     verticalAlignment = Alignment.CenterVertically,