+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Version 0.0.3 :
 
+Feature:
+ - On secret snackbar clicked, navigate to secrets view 
 Fix:
  - Topbar title when fact selected and going back to the screen category
  - Topbar title when random is clicked